+++ changelog.md
@@ -16,6 +16,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 
+- Boxlang Home fixed for windows
 - Console reading via < /dev/tty
 
 ## [1.6.0] - 2025-06-05